// src/app/Medical/Models/applicantmaster.ts
export class ApplicantMaster {
  ID: any = 0;
  CLIENT_ID = 0;
  NAME: any;
  STATUS: boolean = true;
  EMPLOYEE_CODE: any;
  GRADE_PAY: any;
  OFFICE_NAME: any;
  DESIGNATION: any;
  DESIGNATION_ID: any;
  LOCATION: any;
  DDO_OF_THE_OFFICIAL: any;
  BENEFICIARY_TYPE: string = '';
  CGHS_CARD_NO: any;
  CGHS_CARD_VALIDITY: any;
  EMP_ID: any;
  MOBILE_NO: any;
  EMAIL_ID: any;
  ADDRESS: any;
  CURRENT_STAGE_ID: any;
  INSPECTOR_ID: any;
  ROLE_ID: any;
  DDO_OF_THE_OFFICIAL_ID: any;
  IS_APPLYING_FOR_ADVANCE:any;
  SERVICE_TYPE: any;
  IS_ADVANCE_AMOUNT_CLAIMED:any;
}