// src/app/Medical/Models/claimmaster.ts
export class ClaimMaster {
  ID: number;
  // EMP_ID:any;
  // PATIENT_NAME:any;
  // RELATION_WITH_PATIENT:any;
  // NATURE_OF_TREATMENT:any;
  // TREATEMENT_START_DATE    :any;
  // TREATEMENT_END_DATE    :any;
  // IS_ADVANCE_TAKEN:any;
  // ADVANCE_AMOUNT:any;
  ADVANCE_TAKEN_DATE: any;
  CGHS_CARD: any;
  APP_INFO_DATE_TIME: any;
  // INSPECTOR_ID:any;
  CLAIMED_ADVANCE_AMOUNT: any;
  // CURRENT_STAGE_ID:any;
  // NAME:any;
  IS_APPLYING_FOR_ADVANCE: any = 0;
  HOSPITAL_TYPE: string = '';
  // HOSPITAL_ID:any
  HOSPITAL_NAME: any;
  HOSPITAL_ADDRESS: any;
  // BILL_FILLING_DATE:any
  BANK_NAME: any;
  SB_ACC_NO: any;
  HOSPITAL_ORIGINAL_BILL: any;
  SELF_BANK_DETAILS: any;
  IFSC_CODE: any;
  MICR_CODE: any;
  WARD_ENTITLEMENT: string = '';
  TREATEMENT_TYPE: string = '';
  // EMPLOYEE_NAME:any
  EMP_ID: any;
  GRADE_PAY: any;
  OFFICE_NAME: any;
  DESIGNATION: any;
  LOCATION: any;
  DDO_OF_THE_OFFICIAL_ID: any;
  PATIENT_NAME: any;
  RELATION_WITH_PATIENT: any;
  BENEFICIARY_TYPE: any;
  CGHS_CARD_NO: any;
  IS_EMERGENCY_TREATMENT_APPLIED: any;
  EMERGENCY_REF_DATE: any;
  HOSPITAL_ID: any;
  IS_PRIVATE_HOSPITAL: any;
  IS_PERMISSION_TAKEN: any;
  NATURE_OF_TREATMENT: any;
  TREATMENT_START_DATE: any;
  TREATMENT_END_DATE: any;
  IS_FORWARDING_LETTER: any;
  IS_MEDICAL_REIMBURSEMENT_CLAIMED: any;
  IS_COPY_OF_CGHS_CARD_FOR_BOTH_BENEFICIARY_AND_PATIENT: any;
  IS_DISCHARGE_CARD: any;
  FORM_NO_3_IN_CASE_THE_MEDICAL_CLAIM: any;
  IS_FINAL_BILL_IN_ORIGINAL_ALONG_WITH_ONE_XEROX_ONE_COPY: any;
  IS_PAY_SLIP_OF_THE_PERIOD_OF_TRATMENT: any;
  IS_BANK_MANDATE_FROM_ALONG_WITH_CANCELLED_CHEQUE: any;
  IS_PRESCRIPTION_OF_MEDICINES_PURCHASED_FROM_OUTSIDE: any;
  // STAGE_ID: any;
  BILL_FILIING_DATE: any = new Date();
  IS_BILL_FILLED_IN_TIME: any;
  AMOUNT_OF_REIMBUSMENT_CLAIMED: any;
  AMOUNT_OF_REIMBUSMENT_ADMISSIBLE: any;
  IS_ADVANCE_TAKEN: boolean = false;
  ADVANCE_AMOUNT: any;
  NET_AMOUNT_PAYABLE: any;
  CURRENT_STAGE_ID: any;
  INSPECTOR_ID: any;
  REMARK: any;
  SUB_STAGE: any;
  WARD_TYPE: any;
  BANK_ACCOUNT_NO: any;
  CLAIM_ACCREDITATION: any = '';
  PATIENT_CGHS_BENEFICIERY_NO: any;

  ADDRESS: any;
  EMAIL_ID: any;
  STATUS: boolean = true;
  MOBILE_NO: any;
  CGHS_CARD_VALIDITY: any;
  DDO_OF_THE_OFFICIAL: any;
  EMPLOYEE_CODE: any;
  EMPLOYEE_NAME: any;
  PAY_BILL: string = 'NonGazPay';
  hospitalData: any = [];
  DELAY_CONDELENCE_DIFFERENCE: number = 0;
  SIGNATURE_ID: number = 0;
  ADVANCE_ADMISSIBLE_AMOUNT: number = 0;
  BILL_NO: any;
  CLAIM_FORM: any;
  STEP_NO: any;
  CLAIM_INFO_DATE_TIME: any;
  CLAIM_DOWNLOAD_DATE_TIME: any;
  CLAM_UPLOAD_DATE_TIME: any;
  SEND_FOR_VERIFICATION_DATE_TIME: any;
  IS_ADVANCE_AMOUNT_CLAIMED: boolean;
  OPD_TREATEMENT_AMOUNT: any = 0;
  INDOOR_TREATEMENT_AMOUNT: any = 0;
  TEST_OR_INVESTIGATION_AMOUNT: any = 0;
  TOTAL_ADVANCE_CLAIMED_AMOUNT: any = 0;
  ZAO_HN: any;
  ZAO: any;
  CHEQUE_DRAWN: any = 0;
  ESTIMATE: any;
  PAYSLIP: any;
  CGHS_AMA: any;

  HOSPITAL_DISCHARGE_CARD: any;

  IS_ADVANCE_ANNEXURE_CREATED: boolean = false;
  IS_CERTIFICATE_FROM_OFFICER: boolean = false;
  IS_ANTICIPATED_COST_OF_TREATMENT: boolean = false;
  ANTICIPATE_COST_OF_TRATEMENT_CERTIFICATE: any;
  CERTIFICATE_FROM_MEDICAL_OFFICER: any;
  ANTICIPATED_COST: any = 0;
  HOSPITAL_ESTIMATED_AMOUNT: any = 0;
  TENTATIVE_ADMISSION_DATE: any;
  SERVICE_TYPE: any;

  PRECESCRIPTION_ISSUED_BY_HOSPITAL: any;
  OTHER_DOCUMENT: any;
  BANK_DETAILS_HOSPITAL: any;
  TREATEMENT_IS_RECIVED: any;
  SUB_INSPECTOR_ID: any;
  OFFICE_ADDRESS: any;
  PREPARED_BY: any;
  CHECKED_BY: any;
  IS_ADMIN_CLAIM: boolean = true;

  IS_PERMISSION_TAKEN_FROM_HOD_EMP: boolean = false;
  EMERGENCY_TREATEMENT_EMP: boolean = false;
}